+++ naz/codec.py
@@ -33,8 +33,8 @@
 
 
 import sys
+import abc
 import codecs
-import typing
 
 
 # An alternative to using this codec module is to use: https://github.com/dsch/gsm0338
@@ -172,9 +172,51 @@ class UCS2Codec(codecs.Codec):
         return codecs.utf_16_be_decode(input, errors)  # pytype: disable=module-attr
 
 
-class SimpleCodec(codecs.Codec):
+class BaseCodec(abc.ABC):
     """
-    This is an implementation of `codecs.Codec <https://docs.python.org/3/library/codecs.html>`_
+    This is the interface that must be implemented to satisfy naz's encoding/decoding.
+    User implementations should inherit this class and
+    implement the :func:`__init__ <BaseCodec.__init__>`, :func:`encode <BaseCodec.encode>` and :func:`decode <BaseCodec.decode>`
+    methods with the type signatures shown.
+
+    naz calls an implementation of this class to encode/decode messages.
+    """
+
+    @abc.abstractmethod
+    def __init__(self, encoding: str, errors: str) -> None:
+        """
+        Parameters:
+            encoding: `encoding <https://docs.python.org/3/library/codecs.html#standard-encodings>`_ used to encode messages been sent to SMSC
+                      The encoding should be one of the encodings recognised by the SMPP specification. See section 5.2.19 of SMPP spec
+                      eg gsm0338, ucs2 etc
+            errors:	same meaning as the errors argument to pythons' `encode <https://docs.python.org/3/library/codecs.html#codecs.encode>`_ method
+        """
+        self.encoding = encoding
+        self.errors = errors
+
+    @abc.abstractmethod
+    def encode(self, input: str) -> bytes:
+        """
+        return an encoded version of the string as a bytes object
+
+        Parameters:
+            input: the string to encode
+        """
+        raise NotImplementedError("encode method must be implemented.")
+
+    def decode(self, input: bytes) -> str:
+        """
+        return a string decoded from the given bytes.
+
+        Parameters:
+            input: the bytes to decode
+        """
+        raise NotImplementedError("decode method must be implemented.")
+
+
+class SimpleCodec(BaseCodec):
+    """
+    This is an implementation of `BaseCodec`
 
     SMPP uses a 7-bit GSM character set. This class implements that encoding/decoding scheme.
     This class can also be used with the usual `python standard encodings <https://docs.python.org/3/library/codecs.html#standard-encodings>`_
@@ -192,19 +234,25 @@ class SimpleCodec(codecs.Codec):
 
     custom_codecs = {"gsm0338": GSM7BitCodec(), "ucs2": UCS2Codec()}
 
-    def __init__(self, encoding: str = "gsm0338") -> None:
+    def __init__(self, encoding: str = "gsm0338", errors: str = "strict") -> None:
         """
         Parameters:
             encoding: `encoding <https://docs.python.org/3/library/codecs.html#standard-encodings>`_ used to encode messages been sent to SMSC
                       The encoding should be one of the encodings recognised by the SMPP specification. See section 5.2.19 of SMPP spec
+            errors:	same meaning as the errors argument to pythons' `encode <https://docs.python.org/3/library/codecs.html#codecs.encode>`_ method
         """
         if not isinstance(encoding, str):
             raise ValueError(
                 "`encoding` should be of type:: `str` You entered: {0}".format(type(encoding))
             )
+        if not isinstance(errors, str):
+            raise ValueError(
+                "`errors` should be of type:: `str` You entered: {0}".format(type(errors))
+            )
         self.encoding = encoding
+        self.errors = errors
 
-    def encode(self, input: str, errors: str = "strict") -> typing.Tuple[bytes, int]:
+    def encode(self, input: str) -> bytes:
         if not isinstance(input, str):
             raise NazCodecException("Only strings accepted for encoding.")
         encoding = self.encoding or sys.getdefaultencoding()
@@ -212,9 +260,9 @@ class SimpleCodec(codecs.Codec):
             encoder = self.custom_codecs[encoding].encode
         else:
             encoder = codecs.getencoder(encoding)
-        return encoder(input, errors)
+        return encoder(input, self.errors)[0]
 
-    def decode(self, input: bytes, errors: str = "strict") -> typing.Tuple[str, int]:
+    def decode(self, input: bytes) -> str:
         if not isinstance(input, (bytes, bytearray)):
             raise NazCodecException("Only bytestrings accepted for decoding.")
         encoding = self.encoding or sys.getdefaultencoding()
@@ -222,4 +270,4 @@ class SimpleCodec(codecs.Codec):
             decoder = self.custom_codecs[encoding].decode
         else:
             decoder = codecs.getdecoder(encoding)
-        return decoder(input, errors)
+        return decoder(input, self.errors)[0]