+++ CHANGELOG.md
@@ -4,7 +4,7 @@ most recent version is listed first.
 
 ## **version:** v0.7.5
 - BugFix, `data_coding` should not be applied to all fields that are of type C-octet string: https://github.com/komuw/naz/pull/190   
-- Make `naz.Client.codec_class` accept any child class implementation of python's `codecs.Codec`: https://github.com/komuw/naz/pull/190  
+- Rename `naz.nazcodec.BaseNazCodec` to `naz.codec.BaseCodec`: https://github.com/komuw/naz/pull/190  
 - Rename `naz.Client.codec_class` to `naz.Client.codec`: https://github.com/komuw/naz/pull/190